test extraction without linking
