+++ tests/conftest.py
@@ -4,6 +4,7 @@ from pathlib import Path
 import pytest
 
 from deet.processors.converter_register import SupportedImportFormat
+from deet.processors.parser import ParsedOutput
 
 
 @pytest.fixture
@@ -22,6 +23,35 @@ def valid_parsed_epub():
 def valid_parsed_html():
     with Path.open("tests/test_files/output/conrad-html-parsed.md") as infile:
         return infile.read()
+
+
+@pytest.fixture
+def mock_check_language(monkeypatch):
+    """Stub the language checker."""
+    monkeypatch.setattr(
+        "deet.processors.parser.check_language",
+        lambda txt, lang=None, threshold=0.2: txt.strip() != "not english",  # noqa: ARG005
+    )
+
+
+@pytest.fixture
+def mock_pdfminerparser_parse(monkeypatch):
+    """Stub PdfminerParser.parse to avoid actual PDF parsing."""
+
+    def _stub_parse(
+        cls,
+        input_,
+        *,
+        return_metadata: bool = False,
+        return_images: bool = False,
+        **kwargs,
+    ) -> ParsedOutput:
+        return ParsedOutput(text="dummy pdfminer text", parser_library="pdfminer")
+
+    monkeypatch.setattr(
+        "deet.processors.parser.PdfminerParser.parse",
+        classmethod(_stub_parse),
+    )
 
 
 @pytest.fixture
